+++ docs/assets/js/common.d4eda587.js
@@ -1,0 +1,2 @@
+console.log(">> Global loaded script.");
+//# sourceMappingURL=common.d4eda587.js.map