+++ examples/hello-world-app/src/views/pages/home/main.js
@@ -1,0 +1,6 @@
+import App from 'App/app';
+
+window.addEventListener('load', function (event) {
+  console.log('Start app!');
+  App.run();
+});