+++ examples/webpack-app-hello-pug/src/assets/scripts/main.js
@@ -1,6 +1,0 @@
-import { app } from 'App/app';
-
-window.addEventListener('load', function (event) {
-  console.log('Start app!');
-  app();
-});