+++ src/exeptions.js
@@ -17,7 +17,7 @@ class PugLoaderException extends Error {
  * @param {PugLoaderException|Error|string?} error The original error from catch()
  * @constructor
  */
-const PugLoaderError = function(message, error = '') {
+const PugLoaderError = function (message, error = '') {
   if (error && error instanceof PugLoaderException) {
     if (error.toString() === lastError) {
       // prevent double output same error
@@ -39,9 +39,8 @@ const PugLoaderError = function(message, error = '') {
  */
 const resolveException = (error, file, templateFile) => {
   const message =
-    `${ansisLoaderName} the file ${ansis.yellow(
-      file,
-    )} can't be resolved in the pug template:\n` + ansis.cyan(templateFile);
+    `${ansisLoaderName} the file ${ansis.yellow(file)} can't be resolved in the pug template:\n` +
+    ansis.cyan(templateFile);
 
   PugLoaderError(message, error);
 };
@@ -54,10 +53,10 @@ const resolveException = (error, file, templateFile) => {
 const unsupportedInterpolationException = (value, templateFile) => {
   const message =
     `${ansisLoaderName} the expression ${ansis.yellow(
-      value,
+      value
     )} can't be interpolated with the 'compile' method in the pug template: ${ansis.cyan(templateFile)}\n` +
     `${ansis.yellow(
-      'Possible solution: ',
+      'Possible solution: '
     )} Try to use the loader option 'method' as 'render' or change your dynamic filename to static or use webpack alias instead of alias from tsconfig.`;
 
   PugLoaderError(message);
@@ -76,7 +75,7 @@ const executeTemplateFunctionException = (error, sourceFile) => {
     `${ansis.black.bgYellow(`Solution`)} in this case pass a variable into a pug file via the query parameter.\n` +
     `For example, if in pug is used the external variable, like ${ansis.yellow(`title= customData.options.title`)},\n` +
     `then pass it into pug ${ansis.magenta(
-      `'template.pug?customData=' + JSON.stringify({options:{title:'My title'}})`,
+      `'template.pug?customData=' + JSON.stringify({options:{title:'My title'}})`
     )}`;
 
   PugLoaderError(message, error);
@@ -88,9 +87,33 @@ const executeTemplateFunctionException = (error, sourceFile) => {
  * @throws {Error}
  */
 const filterNotFoundException = (filterName, availableFilters) => {
-  const message = `${ansisLoaderName} The 'embedFilters' option contains unknown filter name: ${ansis.red(
-      filterName)}.\n` +
+  const message =
+    `${ansisLoaderName} The 'embedFilters' option contains unknown filter: ${ansis.red(filterName)}.\n` +
     `Available embedded filters: ${ansis.green(availableFilters)}.`;
+
+  PugLoaderError(message);
+};
+
+/**
+ * @param {string} filterName
+ * @param {string} filterPath
+ * @param {Error} error
+ */
+const filterLoadException = (filterName, filterPath, error) => {
+  const message =
+    `${ansisLoaderName} Error by load the ${ansis.red(filterName)} filter.\n` +
+    `Filter file: ${ansis.cyan(filterPath)}\n` +
+    error;
+
+  PugLoaderError(message);
+};
+
+/**
+ * @param {string} filterName
+ * @param {Error} error
+ */
+const filterInitException = (filterName, error) => {
+  const message = `${ansisLoaderName} Error by initialisation the ${ansis.red(filterName)} filter.\n` + error;
 
   PugLoaderError(message);
 };
@@ -108,6 +131,8 @@ module.exports = {
   resolveException,
   unsupportedInterpolationException,
   executeTemplateFunctionException,
-  filterNotFoundException,
   getPugCompileErrorMessage,
+  filterNotFoundException,
+  filterLoadException,
+  filterInitException,
 };