+++ test/cases/exception-filter-not-found/webpack.config.js
@@ -29,14 +29,6 @@ module.exports = {
           },
         },
       },
-
-      {
-        test: /\.(png|jpg|jpeg)/,
-        type: 'asset/resource',
-        generator: {
-          filename: 'assets/images/[name].[hash:8][ext]',
-        },
-      },
     ],
   },
 };